backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
databases==0.9.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
PyMySQL==1.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
